Testing

+++ view_houses_rent.php
@@ -18,16 +18,24 @@ if (isset($_POST['favourite'])) {
 
 if (isset($_POST['unfavourite'])) {
   $dbid = strip_tags($_POST['hid']);
-  $sql = "DELETE FROM favorites WHERE house_id= '$dbid'";
+  $sql = "DELETE FROM favorites WHERE house_id= '$dbid' AND user_id='$userid' ";
   mysqli_query($conn, $sql);
   header("Location: view_houses_rent.php");
 }
 
-if (isset($_POST['delete'])) {
+if (isset($_POST['compare'])) {
   $dbid = strip_tags($_POST['hid']);
-  $sql = "DELETE FROM approval_land WHERE id= '$dbid'";
+  $sql = "INSERT INTO compare_list(user_id,house_id)
+        VALUES ('$userid','$dbid')";
   mysqli_query($conn, $sql);
-  header("Location: view_houses.php");
+  header("Location: view_houses_rent.php");
+}
+
+if (isset($_POST['uncompare'])) {
+  $dbid = strip_tags($_POST['hid']);
+  $sql = "DELETE FROM compare_list WHERE house_id= '$dbid' AND user_id='$userid' ";
+  mysqli_query($conn, $sql);
+  header("Location: view_houses_rent.php");
 }
 
 ?>
@@ -164,9 +172,22 @@ if (isset($_POST['delete'])) {
                         </div>
 
                       <?php } ?>
+                      <?php
+                      $h_id = $row['id'];
+                      $sql = "SELECT house_id FROM compare_list where user_id='$userid' AND house_id='$h_id'";
+                      $result1 = mysqli_query($conn, $sql);
+                      $row1 = mysqli_fetch_array($result1);
+                      if ($row1 == NULL) {
+                      ?>
                       <div class="col-md-6">
                         <button type="submit" name="compare" class="btn btn-danger btn-block"> Add to compare </button>
                       </div>
+                      <?php } else {
+                      ?>
+                      <div class="col-md-6">
+                        <button type="submit" name="uncompare" class="btn btn-danger btn-block"> <i class="fa fa-check-square-o" aria-hidden="true"></i> Compared already </button>
+                      </div>
+                      <?php } ?>
                     </div>
                   </form>
                 </div>
